Datos de ejemplo, censo 2017 || || Example data, 2017 census (Chile)

Data de ejemplo proveniente del censo 2017. Corresponde a la comuna de San Pablo, X Región de Los Lagos. Se eligió por su pequeño tamaño, adecuado para límites de tamaño de archivo de CRAN y GitHub. || || Example data from 2017 Chilean census. It is the data from San Pablo commune. It was chosen because of its small size, appropriate for CRAN and GitHub file size limitations.

c17_example

Format

## `c17_example` A data frame with 7512 rows and 60 columns:

region

ID de la región || Region ID

provincia

ID de la provincia || Province ID

comuna

ID de la comuna || Commune ID

dc

ID del distrito || District ID

area

ID del área || Area ID

zc_loc

Zona local || Local zone

id_zc_loc

ID de zona local || Local zone ID

nviv

Número de la vivienda || Dwelling number

p01 - p16a

Preguntas del censo || Census questions