Limpieza de la base de datos censal -- Census database cleanup

Limpia la base de datos de forma de normalizar los nombres de los campos y reducir la cantidad de variables, facilitando así la ejecución de las funciones que siguen en el flujo de cálculo.

Cleans the database, normalizing the field names and reducing the number of variables, facilitiating the execution of the following functions down the workflow.

cleanup(
  df,
  year = 2017,
  tiphog = "p01",
  ocupac = "p02",
  ndorms = "p04",
  parent = "p07",
  muro = "p03a",
  techo = "p03b",
  suelo = "p03c"
)

Arguments

df

objeto data.frame. Se recomienda usar la base original del Censo 2017, disponible a través de load_data().

data.frame object; recommendation is to use the original 2017 census database, available through load_data().

year

integer. Default es 2017, que es la única disponible para la versión 1.x.x.

integer. Default is 2017, which is the only supported for version 1.x.x.

tiphog

string. Nombre del campo de tipo de hogar. Default es p01.

string. Name of the home type field. Default is p01.

ocupac

string. Nombre del campo de ocupación de la vivienda. Default es p02.

string. Name of the home occupation field. Default is p02.

ndorms

string. Nombre del campo con el número de dormitorios del hogar. Default es p04.

string. Name of the number of bedrooms field. Default is p04.

parent

string. Nombre del campo de parentesco. Default es p07.

string. Name of the familial relationship field. Default is p07.

muro

string. Nombre del campo de condición del muro. Default es p03a.

string. Name of the wall condition field. Default is p03a.

techo

string. Nombre del campo de condición del techo. Default es p03b.

string. Name of the ceiling condition field. Default is p03b.

suelo

string. Nombre del campo de condición del suelo. Default es p03c.

string. Name of the floor condition field. Default is p03c.

Value

objeto data.frame conteniendo solo las variables necesarias para los cálculos siguientes.

data.frame object containing only the variables that are necessary for the following calculations.

Examples

c17 <- load_data(13, path = loc_dir) %>% region_filter(13, 1) %>% cleanup()
#> Error in load_data(13, path = loc_dir) %>% region_filter(13, 1) %>% cleanup(): could not find function "%>%"