Geofilter: filtro en base a datos geográficos || Geofilter: filter based in geographical data

Geofilter: filtro en base a datos geográficos || Geofilter: filter based in geographical data

geofilter(df, year, r, area, rfield = "id_region", urfield = "tipo_area")

Arguments

df

Objeto data.frame corresponiente a la base de datos censales del año especificado.
data.frame object corresponding to the census database of the specified year.

year

int. Año de la base de datos. Para la versión 2.x.x, acepta las bases de los censos 2017, 2012, 2002, 1992 y 1982. Default es 2017.
int. Year of the database. For version 2.x.x, census bases for years 2017, 2012, 2002, 1992 and 1982 are accepted.

r

int. Número de la región de trabajo. Para los años 2017 y 2012 se aceptan valores en range(1 : 16, 99), y para los años 2002, 1992 y 1982 se aceptan valores en range(1 : 15, 99).
int. Number od the working region. For years 2017 and 2012 values in range(1 : 16, 99) are accepted, while for years 2002, 1992 and 1982 values in range(1 : 15, 99).

area

int. Código del tipo de área: 1 para URBANO o 2 para RURAL. Adicionalmente se da la opción 3 para TODO. A nivel de data, la categoría RURAL es válida para todos los censos; sin embargo, se debe considerar para posteriores análisis que solo 2017 y 2012 cuentan con coberturas geográficas para las áreas rurales.
int. Area type code: 1 for URBANO or 2 for RURAL. Aditionally, the option 3 for TODO is provided. At a data level, the RURAL category is valid for all censusses; however, it must be considerered for futher analysis that only 2017 and 2012 have spatial data for rural areas.

rfield

int. Nombre de la variable que contiene el código de región. Default es id_region.
int. Name of the variable that contains the region code.

urfield

int. Nombre de la variable que contiene el tipo de área. Default es tipo_area. int. Name of the variable that contains the area type. Default is tipo_area.

Value

Objeto data.frame con la data filtrada.

data.frame object with the filtered data.

Examples

c17rx <- region_filter(c17, 13, 1)
#> Error in region_filter(c17, 13, 1): could not find function "region_filter"