Espacializando data del ISMT -- Spacializing ISMT data

Espacializa la data obtenida del cálculo del ISMT mediante la unión con un shapefile de la unidad espacial correspondiente.

Spacializes the data obtained from ISMT caclulation through merging with a shapefile of the corresponding spatial unit.

geomerge(df, shp, grouping.df = "zona", grouping.shp = "GEOCODIGO")

Arguments

df

objeto data.frame. Asume que los datos estarán agrupados a la unidad espacial correspondiente al shapefile especificado, tras la ejecución de ismt_scores().

shp

objeto simple feature (ESRI shapefile). El uso esperado incluye el shapefile censal, disponible en la página oficial del Censo o a través de load_shp() para zonas urbanas y localidades rurales.

grouping.df

string. Nombre del campo con la variable agrupadora en la tabla de datos

string. Name of the field containing the grouping variable in the data table.

grouping.shp

string. Nombre del campo con la variable agrupadora en el shapefile.

string. Name of the field containing the grouping variable in the shapefile.

Value

objeto data.frame con la información espacializada, lista para exportar.

data.frame object containing the spacialized information, ready for export.

Examples

c17geomerge <- geomerge(c17, c17geo)
#> Error in geomerge(c17, c17geo): object 'c17' not found