Cálculos finales de ISMT -- ISMT final calculations

Ejecuta los cálculos finales de ISMT. Define los grupos socioeconómicos por unidad territorial y los cuantifica en varias categorías.

Executes the final ISMT calculations. Defines the socio-economic groups and quantifies them in several categories.

ismt_scores(df, r, ismt_score = "ismt_pn", grouping = "geocode")

Arguments

df

objeto data.frame. Asume que la base de datos ha pasado por cleanup(), precalc(), y get_pca().

data.frame object. Assumes the database has been through cleanup(), precalc(), and get_pca().

r

integer. Número de la región de trabajo. Acepta valores entre 1 y 16; si r = 99, se utilizan valores a nivel nacional.

integer. Number of the working region. Accepts values between 1 and 16; if r = 99, national-level values will be used.

ismt_score

string. Nombre del campo del puntaje ISMT, calculado desde get_pca(). Default es ismt_pn.

string. Name of the ISMT sscore field, as calculated from get_pca()]. Default is ismt_pn.

grouping

string. Nombre del campo con la variable de la unidad espacial agrupadora. Default es geocode.

string. Name of the field with the spacial grouping unit variable. Default is geocode.

criteria

objeto data.frame que contenga los criterios estadísticos AIM para el análisis de componentes principales.

data.frame object that contains the AIM statistical criteria for principal components analysis.

n

string. Nombre del campo derivado de cleanup() que tiene valor n = 1 para cada jefe de hogar. Usado para conteo de totales. Default es n.

string. Name of the field, derived from cleanup(), that has value n = 1 for each home head. Default es n.

Value

objeto data.frame agrupado por la unidad espacial especificada con información de ISMT.

data.frame object grouped by the specified spatial unit with ISMT information.

Examples

c17 <- load_data(13, path = loc_dir) %>% region_filter(13, 1) %>% cleanup() %>% precalc() %>% get_pca() %>% ismt_scores(crit_AIM, 13)
#> Error in load_data(13, path = loc_dir) %>% region_filter(13, 1) %>% cleanup() %>%     precalc() %>% get_pca() %>% ismt_scores(crit_AIM, 13): could not find function "%>%"