Literalizar bases censales -- Literalize census databases

Literalizar bases censales -- Literalize census databases

literalize(df, year)

Arguments

df

objeto data.frame correspondiente a la base de datos original de los censos 2017, 2012, 2002, 1992 o 1982. -- data.frame object corresponding to the original database for the 2017, 2012, 2002, 1992 or 1982 census.

year

integer. No provee Default para forzar la explicitación del año de la base.

integer. Default is not provided, so as to force explicitation of the database's year.

Value

objeto data.frame con las variables como texto en lugar de factores (nombres y categorías homologados para todos los censos)

Examples

# lit <- c17 |> literalize(2017)