Descargar data censal -- Downloading census data

Descarga la base de datos original del Censo 2017 de Chile a path y la carga a RStudio, quedando lista para su uso.

Downloads the original 2017 Chilean Census database to path and loads it to RStudio, making it immediately available for use.

load_data(r, year = 2017, path = "")

Arguments

r

integer. Número de la región de trabajo. Acepta valores entre 1 y 16. Si r = 99, descarga la data censal completa. r = 99 y r = 13 aumentan considerablemente el tiempo de ejecución debido a la gran cantidad de datos.

integer. Number of the working region. Accepts values between 1 and 16. If r = 99, will download the entire database. r = 99 and r = 13 considerably increase execution time due to the volume of data.

year

integer. Define el año de los datos a descargar. Para la versión 1.x.x, solo se acepta 2017.

integer. Defines the year of the data to download. As for version 1.x.x, only 2017 is accepted.

path

string. Ubicación de la carpeta donde se deberían guardar los archivos. En caso de no proveer una ubicación, se creará un carpeta C:/ISMT_Files/data, que se usará como path.

string. Location of the directory where files should be saved. If empty, will create a C:/ISMT_Files/data folder, which will be used as path.

Value

data.frame object

Examples

c17 <- load_data(13, path = loc_dir)
#> Warning: 'C:\temp' already exists
#> Error in load_data(13, path = loc_dir): object 'loc_dir' not found