Precálculos ISMT -- ISMT Precalculations

Ejecuta precálculos necesarios para el resto del proceso del ISMT.

Executes precalculations necessary for the rest of the ISMT process.

precalc(
  df,
  hacin = "ind_hacinam",
  alleg = "n_hog_alleg",
  esc = "a_esc_cont",
  mat = "ind_mater"
)

Arguments

df

objecto data.frame. Asume que la base de datos ha pasado por cleanup().

data.frame object. Assumes the database has been through cleanup().

hacin

string. Nombre del campo del indicador de hacinamiento. Default es ind_hacinam.

string. Name of the field with the overcrowding score. Default is ind_hacinam.

alleg

string. Nombre del campo del indicador de allegamiento. Default es n_hog_alleg.

string. Name of the field with the relative crowding score. Default is n_hog_alleg.

esc

string. Nombre del campo del indicador de escolaridad del jefe de hogar. Default es a_esc_cont.

string. Name of the field with the scholarship of the home head score. Default is a_esc_cont.

mat

string. Nombre del campo del indicador de materialidad de la vivienda. Default es ind_mater.

string. Name of the field with the dwelling material score. Default is ind_mater.

Value

objeto data.frame con los precálculos necesarios para calcular el ISMT.

data.frame object with the necessary precalculations to calculate ISMT.

Examples

c17 <- load_data(13, path = loc_dir) %>% region_filter(13, 1) %>% cleanup() %>% precalc()
#> Error in load_data(13, path = loc_dir) %>% region_filter(13, 1) %>% cleanup() %>%     precalc(): could not find function "%>%"