Filtro de datos por región y área -- Filter data by region and area type

Filtra la base de datos original de forma de incluir sólo la región y el área (urbana o rural) deseadas. La función asume que la data corresponde a la base de dato original del Censo 2017, obtenible a través de load_data().

Filters the original data so as to include only the desired region and area (urban or rural). The function assumes the data corresponds to the original Censo 2017 database, as can be obtained from load_data().

region_filter(df, r, ur, rfield = "region", urfield = "area")

Arguments

df

objeto data.frame. Hasta la versión 1.x.x, sólo se acepta el Censo 2017.

data.frame object. As for package version 1.x.x, only the 2017 Chilean census is accepted.

r

integer. Se aceptan valores entre 1 y 16 para Chile 2017. Si r == 99, no se define una región en particular y se trabaja con valores a nivel nacional.

integer. Values between 1 and 16 are acceptable for Chile 2017. If r == 99, no region will be defined and work will continue with national level values.

ur

integer. Valores aceptables son 1 y 2. Define si se requiere zona urbana ur = 1 o rural ur = 2.

integer. Accepted values are 1 and 2. Defines whether urban ur = 1 or rural ur = 2 data is requested.

rfield

string. Nombre del campo que corresponde al número de la región. Default es region.

string. Name of the field corresponding to the region number. Default is region.

urfield

string. Nombre del campo que define el tipo de área (urbana o rural). Default es area.

string Name of the field corresponding to the desired area (urban or rural). Default is area.

Value

objeto data.frame con la información censal filtrada para la región de trabajo.

data.frame object with the census information filtered for the working region.

Examples

c17 <- load_data(13, path = loc_dir) %>% region_filter(13, 1)
#> Error in load_data(13, path = loc_dir) %>% region_filter(13, 1): could not find function "%>%"