Limpieza de la base de datos censal -- Census database cleanup

Limpia la base de datos de forma de normalizar los nombres de los campos y reducir la cantidad de variables, facilitando así la ejecución de las funciones que siguen en el flujo de cálculo. || || Cleans the database, normalizing the field names and reducing the number of variables, facilitiating the execution of the following functions down the workflow.

cleanup(
  df,
  year = 2017,
  vars.as.factors = FALSE,
  level = "zc",
  tipo_viv = "tipoviv",
  ocupacion = "ocup_viv",
  parentesco = "parentesco",
  dormitorios = "ndorms",
  muro = "mat_muro",
  techo = "mat_techo",
  piso = "mat_piso"
)

Arguments

df

objeto data.frame. Se recomienda usar la base original del Censo 2017, disponible a través de la página oficial del INE. || || data.frame object; recommendation is to use the original 2017 census database, available through INE's official website.

year

integer. Default es 2017. || || integer. Default is 2017.

vars.as.factors

boolean. Si las variables están como factores (como en las bases censales originales), cambiar a TRUE. Default es FALSE, ya que se asume que la base pasá por la función literalize(). || || boolean. If variables are as factors (like in the original census databases), change to TRUE. Default is FALSE, as it is assumed the database has been through the literalize() function.

level

string. Nivel de agrupación de los datos finales. Acepta valores zc (zona censal) y mzn (manzana). El nivel de manzana no está disponible para el 2017 debido al secreto estadístico de la base de datos de origen. Default es zc. || || string. Grouping level for the final data. Values zc (census zone) and mzn (block). Block level is not available for 2017 due to statistical secret of original database. Default is zc.

tipo_viv

string. Nombre del campo de tipo de vivienda. Default es tipoviv. || || string. Name of the dwelling type field. Default is tipoviv.

ocupacion

string. Nombre del campo de ocupación de la vivienda. Default es ocup_viv. || || string. Name of the home occupation field. Default is ocup_viv.

parentesco

string. Nombre del campo de parentesco. Default es parentesco. || || string. Name of the familial relationship field. Default is parentesco.

dormitorios

string. Nombre del campo con el número de dormitorios del hogar. Default es ndorms. || || string. Name of the number of bedrooms field. Default is ndorms.

muro

string. Nombre del campo de condición del muro. Default es mat_muro. || || string. Name of the wall condition field. Default is mat_muro.

techo

string. Nombre del campo de condición del techo. Default es mat_techo. || || string. Name of the ceiling condition field. Default is mat_techo.

piso

string. Nombre del campo de condición del suelo. Default es mat_piso. || || string. Name of the floor condition field. Default is mat_piso.

Value

objeto data.frame conteniendo solo las variables necesarias para los cálculos siguientes. || || data.frame object containing only the variables that are necessary for the following calculations.

Examples

 data(c17_example)
 clean <- c17_example |> literalize(2017) |> cleanup()