Cálculo de ISMT completo a partir del Censo 2017 -- Full ISMT calculation from 2017 Census.

Ejecuta el cálculo completo del ISMT

full_ismt(
  df,
  r,
  ur,
  rfield = "region",
  urfield = "area",
  year = 2017,
  tiphog = "p01",
  ocupac = "p02",
  ndorms = "p04",
  parent = "p07",
  muro = "p03a",
  techo = "p03b",
  suelo = "p03c",
  grouping = "geocode"
)

Arguments

df

objeto data.frame. Hasta la versión 1.x.x, sólo se acepta el Censo 2017.

data.frame object. As for package version 1.x.x, only the 2017 Chilean census is accepted.

r

integer. Se aceptan valores entre 1 y 16 para Chile 2017. Si r == 99, no se define una región en particular y se trabaja con valores a nivel nacional.

integer. Values between 1 and 16 are acceptable for Chile 2017. If r == 99, no region will be defined and work will continue with national level values.

ur

integer. Valores aceptables son 1 y 2. Define si se requiere zona urbana ur = 1 o rural ur = 2.

integer. Accepted values are 1 and 2. Defines whether urban ur = 1 or rural ur = 2 data is requested.

rfield

string. Nombre del campo que corresponde al número de la región. Default es region.

string. Name of the field corresponding to the region number. Default is region.

urfield

string. Nombre del campo que define el tipo de área (urbana o rural). Default es area.

string Name of the field corresponding to the desired area (urban or rural). Default is area.

year

integer. Default es 2017, que es la única disponible para la versión 1.x.x.

integer. Default is 2017, which is the only supported for version 1.x.x.

tiphog

string. Nombre del campo de tipo de hogar. Default es p01.

string. Name of the home type field. Default is p01.

ocupac

string. Nombre del campo de ocupación de la vivienda. Default es p02.

string. Name of the home occupation field. Default is p02.

ndorms

string. Nombre del campo con el número de dormitorios del hogar. Default es p04.

string. Name of the number of bedrooms field. Default is p04.

parent

string. Nombre del campo de parentesco. Default es p07.

string. Name of the familial relationship field. Default is p07.

muro

string. Nombre del campo de condición del muro. Default es p03a.

string. Name of the wall condition field. Default is p03a.

techo

string. Nombre del campo de condición del techo. Default es p03b.

string. Name of the ceiling condition field. Default is p03b.

suelo

string. Nombre del campo de condición del suelo. Default es p03c.

string. Name of the floor condition field. Default is p03c.

grouping

string. Nombre del campo con la variable de la unidad espacial agrupadora. Default es geocode.

string. Name of the field with the spacial grouping unit variable. Default is geocode.

Value

objeto data.frame conteniendo

Examples

ismt <- full_ismt(c17, 13, 1)
#> Error in ismtchile::region_filter(., r = r, ur = ur, rfield = rfield,     urfield = urfield): object 'c17' not found