Cálculo de ISMT completo a partir del Censo 2017 -- Full ISMT calculation from 2017 Census.

Ejecuta el cálculo completo del ISMT. Agrupa las funciones `literalize()`, `geofilter()`, `cleanup()`, `precalc()`, `get_pca()` e `ismt_scores()`. La tabla obtenida debería ser idéntica al resultado de concatenar estas funciones una a una.

full_ismt(
  df,
  r,
  ur,
  rfield = "id_region",
  urfield = "tipo_area",
  year = 2017,
  tipo_vivienda = "tipoviv",
  ocupacion = "ocup_viv",
  ndorms = "ndorms",
  parentesco = "parentesco",
  muro = "mat_muro",
  techo = "mat_techo",
  piso = "mat_piso",
  grouping = "id_zona",
  level = "zc"
)

Arguments

df

objeto data.frame. Default es 2017.

data.frame object. Default is 2017.

r

integer. Se aceptan valores entre 1 y 16 para Chile 2017. Si r == 99, no se define una región en particular y se trabaja con valores a nivel nacional.

integer. Values between 1 and 16 are acceptable for Chile 2017. If r == 99, no region will be defined and work will continue with national level values.

ur

integer. Valores aceptables son 1 y 2. Define si se requiere zona urbana ur = 1 o rural ur = 2.

integer. Accepted values are 1 and 2. Defines whether urban ur = 1 or rural ur = 2 data is requested.

rfield

string. Nombre del campo que corresponde al número de la región. Default es region.

string. Name of the field corresponding to the region number. Default is region.

urfield

string. Nombre del campo que define el tipo de área (urbana o rural). Default es tipo_area.

string Name of the field corresponding to the desired area (urban or rural). Default is tipo_area.

year

integer. Default es 2017.

integer. Default is 2017.

tipo_vivienda

string. Nombre del campo de tipo de vivienda. Default es tipoviv.

string. Name of the dwelling type field. Default is tipoviv.

ocupacion

string. Nombre del campo de ocupación de la vivienda. Default es ocup_viv.

string. Name of the home occupation field. Default is ocup_viv.

ndorms

string. Nombre del campo con el número de dormitorios del hogar. Default es ndorms.

string. Name of the number of bedrooms field. Default is ndorms.

parentesco

string. Nombre del campo de parentesco. Default es parentesco.

string. Name of the familial relationship field. Default is parentesco.

muro

string. Nombre del campo de condición del muro. Default es mat_muro.

string. Name of the wall condition field. Default is mat_muro.

techo

string. Nombre del campo de condición del techo. Default es mat_techo.

string. Name of the ceiling condition field. Default is mat_techo.

piso

string. Nombre del campo de condición del suelo. Default es mat_piso.

string. Name of the floor condition field. Default is mat_piso.

grouping

string. Nombre del campo con la variable de la unidad espacial agrupadora. Default es id_zona.

string. Name of the field with the spacial grouping unit variable. Default is id_zona.

level

string. Nivel de agrupación de los datos finales. Acepta valores zc (zona censal) y mzn (manzana). El nivel de manzana no está disponible para el 2017 debido al secreto estadístico de la base de datos de origen. Default es zc.

string. Grouping level for the final data. Values zc (census zone) and mzn (block). Block level is not available for 2017 due to statistical secret of original database. Default is zc.

Value

objeto data.frame conteniendo

Examples

ismt <- full_ismt(c17, 13, 1)
#> Error in mutate(., id_region = str_pad(region, width = 2, side = "left",     pad = "0"), id_provin = str_pad(provincia, width = 3, side = "left",     pad = "0"), id_comuna = str_pad(comuna, width = 5, side = "left",     pad = "0"), id_distri = paste(id_comuna, str_pad(dc, width = 2,     side = "left", pad = "0"), sep = ""), id_area = paste(id_distri,     as.character(area), sep = ""), id_zona = paste(id_area, str_pad(zc_loc,     width = 3, side = "left", pad = "0"), sep = ""), id_manzan = NA_character_,     nviv = as.integer(nviv), nhogar = as.integer(nhogar), npers = as.integer(personan),     cant_hog = case_when(cant_hog %in% c(0:35) ~ as.integer(cant_hog),         T ~ NA_integer_), cant_per = as.integer(cant_per), tipoviv = case_when(p01 ==         1 ~ "CASA", p01 == 2 ~ "DPTO", p01 == 3 ~ "INDIGENA",         p01 == 4 ~ "PIEZA", p01 == 5 ~ "MEDIAGUA", p01 == 6 ~             "MOVIL", p01 == 7 ~ "OTRO", p01 == 8 ~ "COLECTIVA",         p01 == 9 ~ "TRANSITO", p01 == 10 ~ "CALLE", T ~ NA_character_),     ocup_viv = case_when(p02 == 1 ~ "PRESENTES", p02 == 2 ~ "AUSENTES",         p02 == 3 ~ "DESOCUPADA", p02 == 4 ~ "TEMPORADA", T ~             NA_character_), mat_muro = case_when(p03a == 1 ~         "HORMIGON", p03a == 2 ~ "ALBANILERIA", p03a == 3 ~ "TABIQUE_FORRADO",         p03a == 4 ~ "TABIQUE", p03a == 5 ~ "ARTESANAL", p03a ==             6 ~ "PRECARIOS", T ~ NA_character_), mat_techo = case_when(p03b ==         1 ~ "TEJAS", p03b == 2 ~ "LOSA_HORMIGON", p03b == 3 ~         "PLANCHAS", p03b == 4 ~ "FONOLITA", p03b == 5 ~ "PAJA",         p03b == 6 ~ "PRECARIOS", p03b == 7 ~ "SIN_TECHO", T ~             NA_character_), mat_piso = case_when(p03c == 1 ~         "PARQUET", p03c == 2 ~ "RADIER", p03c == 3 ~ "BALDOSA_CEMENTO",         p03c == 4 ~ "CEMENTO_TIERRA", p03c == 5 ~ "TIERRA", T ~             NA_character_), ndorms = case_when(p04 %in% c(0:6) ~         as.integer(p04), T ~ NA_integer_), agua_orig = case_when(p05 ==         1 ~ "RED_PUBLICA", p05 == 2 ~ "POZO", p05 == 3 ~ "CAMION",         p05 == 4 ~ "RIO", T ~ NA_character_), tipo_hog = case_when(tipo_hogar ==         1 ~ "UNIPERSONAL", tipo_hogar == 2 ~ "MONOPARENTAL",         tipo_hogar == 3 ~ "BIPARENTAL_NH", tipo_hogar == 4 ~             "BIPARENTAL_CH", tipo_hogar == 5 ~ "COMPUESTO", tipo_hogar ==             6 ~ "EXTENSO", tipo_hogar == 7 ~ "SIN_NUCLEO", T ~             NA_character_), operativo = case_when(tipo_operativo ==         1 ~ "VIV_PARTICULAR", tipo_operativo == 8 ~ "VIV_COLECTIVA",         tipo_operativo == 9 ~ "TRANSITO", tipo_operativo == 10 ~             "CALLE", T ~ NA_character_), parentesco = case_when(p07 ==         1 ~ "JEFE_HOGAR", p07 == 2 ~ "CONYUGUE", p07 == 3 ~ "UNION_CIVIL",         p07 == 4 ~ "PAREJA", p07 == 5 ~ "HIJO_A", p07 == 6 ~             "HIJO_A_CONYUGUE", p07 == 7 ~ "HERMANO_A", p07 ==             8 ~ "PADRE_MADRE", p07 == 9 ~ "CUNADO_A", p07 ==             10 ~ "SUEGRO_A", p07 == 11 ~ "YERNO_NUERA", p07 ==             12 ~ "NIETO_A", p07 == 13 ~ "ABUELO_A", p07 == 14 ~             "OTRO_PARIENTE", p07 == 15 ~ "NO_PARIENTE", p07 ==             16 ~ "SERVICIO", p07 == 17 ~ "VIV_COLECTIVA", p07 ==             18 ~ "OP_TRANSITO", p07 == 19 ~ "OP_CALLE", T ~ NA_character_),     sexo = case_when(p08 == 1 ~ "HOMBRE", p08 == 2 ~ "MUJER",         T ~ NA_character_), edad = as.integer(p09), res_5a = case_when(p11 ==         1 ~ "NO_NACIDO", p11 == 2 ~ "ESTA_COMUNA", p11 == 3 ~         "OTRA_COMUNA", p11 == 4 ~ "PERU", p11 == 5 ~ "ARGENTINA",         p11 == 6 ~ "BOLIVIA", p11 == 7 ~ "ECUADOR", p11 == 8 ~             "COLOMBIA", p11 == 9 ~ "OTRO", p11 %in% c(98, 99) ~             "NO_APLICA", T ~ NA_character_), nacimiento = case_when(p12 ==         1 ~ "ESTA_COMUNA", p12 == 2 ~ "OTRA_COMUNA", p12 == 3 ~         "PERU", p12 == 4 ~ "ARGENTINA", p12 == 5 ~ "BOLIVIA",         p12 == 6 ~ "ECUADOR", p12 == 7 ~ "COLOMBIA", p12 == 8 ~             "OTRO", p12 %in% c(98, 99) ~ "NO_APLICA", T ~ NA_character_),     asiste_educ = case_when(p13 == 1 ~ "SI", p13 == 2 ~ "NO_ACTUAL",         p13 == 3 ~ "NUNCA", T ~ NA_character_), curso_alto = case_when(p14 %in%         c(0:8) ~ as.integer(p14), T ~ NA_integer_), nivel_educ = case_when(p15 ==         1 ~ "JARDIN", p15 == 2 ~ "PREKINDER", p15 == 3 ~ "KINDER",         p15 == 4 ~ "DIFERENCIAL", p15 == 5 ~ "BASICA", p15 ==             6 ~ "PRIMARIA", p15 == 7 ~ "CIENTIFICO_HUMANISTA",         p15 == 8 ~ "TECNICA_PROF", p15 == 9 ~ "HUMANIDADES",         p15 == 10 ~ "TEC_COMERCIAL", p15 == 11 ~ "TEC_SUPERIOR",         p15 == 12 ~ "PROFESIONAL", p15 == 13 ~ "MAGISTER", p15 ==             14 ~ "DOCTORADO", T ~ NA_character_), pueblo_pert = case_when(p16 ==         1 ~ "SI", p16 == 2 ~ "NO", T ~ NA_character_), pueblo_orig = case_when(p16a ==         1 ~ "MAPUCHE", p16a == 2 ~ "AYMARA", p16a == 3 ~ "RAPA_NUI",         p16a == 4 ~ "LICAN_ANTAI", p16a == 5 ~ "QUECHUA", p16a ==             6 ~ "COLLA", p16a == 7 ~ "DIAGUITA", p16a == 8 ~             "KAWESQAR", p16a == 9 ~ "YAGAN", p16a == 10 & p16a_otro ==             3 ~ "LAFKENCHE", p16a == 10 & p16a_otro == 4 ~ "PEHUENCHE",         p16a == 10 & p16a_otro == 5 ~ "HUILLICHE", p16a == 10 &             p16a_otro == 6 ~ "PICUNCHE", p16a == 10 & p16a_otro ==             21 ~ "CHANGO", p16a == 10 & p16a_otro == 22 ~ "CHONO",         p16a == 10 & p16a_otro == 23 ~ "ONA", p16a == 10 & p16a_otro ==             28 ~ "TEHUELCHE", p16a == 10 & p16a_otro == 33 ~             "LATAM", p16a == 10 & p16a_otro == 34 ~ "MUNDO",         p16a == 10 & p16a_otro == 35 ~ "AFRO", p16a == 10 & p16a_otro ==             37 ~ "OTROS", p16a == 10 & p16a_otro == 97 ~ "NO_DECLARA",         T ~ NA_character_), trabajo = case_when(p17 == 1 ~ "PAGADO",         p17 == 2 ~ "SIN_PAGO_FAM", p17 == 3 ~ "VACACIONES", p17 ==             4 ~ "BUSCANDO", p17 == 5 ~ "ESTUDIANDO", p17 == 6 ~             "QUEHACERES", p17 == 7 ~ "JUBILADO", p17 == 8 ~ "OTRO",         T ~ NA_character_), rama = case_when(p18 == "A" ~ "AGRO",         p18 == "B" ~ "MINERIA", p18 == "C" ~ "MANUFACTURA", p18 ==             "D" ~ "SUMINISTRO", p18 == "E" ~ "AGUAS", p18 ==             "F" ~ "CONSTRUCCION", p18 == "G" ~ "COMERCIO", p18 ==             "H" ~ "TRANSPORTE", p18 == "I" ~ "ALOJAMIENTO", p18 ==             "J" ~ "INFORMACION", p18 == "K" ~ "FINANZA", p18 ==             "L" ~ "INMOBILIARIA", p18 == "M" ~ "PROFESIONAL",         p18 == "N" ~ "ADMINISTRACION", p18 == "O" ~ "ADMIN_PUBLICA",         p18 == "P" ~ "EDUCACION", p18 == "Q" ~ "SALUD", p18 ==             "R" ~ "ENTRETENIMIENTO", p18 == "S" ~ "OTROS_SERVICIOS",         p18 == "T" ~ "HOGAR", p18 == "U" ~ "EXTRATERRITORIO",         p18 == "Z" ~ "NO_DECLARA", T ~ NA_character_), hijos_nacido = case_when(p19 %in%         c(0:23) ~ as.integer(p19), T ~ NA_integer_), hijos_vivos = case_when(p20 %in%         c(0:23) ~ as.integer(p20)), escolaridad = case_when(escolaridad %in%         c(0:21) ~ as.integer(escolaridad), T ~ NA_integer_),     year = "2017", tipo_area = case_when(area == 1 ~ "URBANO",         area == 2 ~ "RURAL", T ~ NA_character_)): object 'c17' not found